# static/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: NHSO Revenue Intelligence
  description: |
    REST API for Hospital Information System (HIS) integration with E-Claim data.

    ## Authentication
    All endpoints (except `/health`) require API key authentication via `X-API-Key` header.

    ## Getting Started
    1. Generate API key from the admin panel
    2. Include the API key in request headers: `X-API-Key: your-api-key`
    3. Use the endpoints to fetch claims data and sync reconciliation

    ## Rate Limits
    - 100 requests per minute per API key
    - 1000 requests per hour per API key

    ## Data Sources
    - **REP**: E-Claim reimbursement data (OP/IP claims)
    - **STM**: Statement reconciliation data
    - **SMT**: Smart Money Transfer budget data

  version: 1.0.0
  contact:
    name: AEGIS-X Platform
    url: https://github.com/aegisx-platform/eclaim-rep-download
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5001/api/v1
    description: Development server
  - url: https://{hospital-domain}/api/v1
    description: Production server
    variables:
      hospital-domain:
        default: hospital.example.com
        description: Your hospital's domain
...
security:
  - ApiKeyAuth: []
paths:
...
  /claims/{tran_id}:
    get:
      tags:
        - Claims
      summary: Get single claim
      description: Retrieve a single claim by TRAN_ID
      operationId: getClaimById
      parameters:
        - name: tran_id
          in: path
          required: true
          description: Transaction ID
          schema:
            type: string
            example: "1234567890"
      responses:
        '200':
          description: Claim found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Claim'
        '404':
          description: Claim not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                error: "Claim not found"
                code: "NOT_FOUND"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication. Generate from admin panel.
  schemas:
    Claim:
      type: object
      properties:
        tran_id:
          type: string
          description: Transaction ID (unique identifier)
          example: "1234567890"
        file_id:
          type: integer
          description: Import file ID
          example: 42
        hn:
          type: string
          description: Hospital Number
          example: "12345678"
        pid:
          type: string
          description: Personal ID (13 digits)
          example: "1234567890123"
        dateadm:
          type: string
          format: date
          description: Admission date
          example: "2025-12-15"
        datedsc:
          type: string
          format: date
          description: Discharge date
          example: "2025-12-16"
        an:
          type: string
          description: Admission Number (for IP claims)
          example: "65120001"
          nullable: true
        rep_no:
          type: string
          description: Reimbursement number
          example: "UCS6512000001"
        total_approve:
          type: number
          format: float
          description: Total approved amount (THB)
          example: 2500.00
        hcode:
          type: string
          description: Hospital code (5 digits)
          example: "10670"
        hmain:
          type: string
          description: Main hospital code
          example: "10670"
        his_matched:
          type: boolean
          description: Whether matched with HIS
          example: true
          nullable: true
        his_vn:
          type: string
          description: HIS Visit Number
          example: "6512000001"
          nullable: true
        reconcile_status:
          type: string
          description: Reconciliation status
          enum: [matched, unmatched, pending]
          example: "matched"
          nullable: true
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Invalid parameters"
        code:
          type: string
          example: "INVALID_PARAMS"
  responses:
...
    Unauthorized:
      description: Authentication failed (invalid or missing API key)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Invalid API key"
            code: "UNAUTHORIZED"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Internal server error"
            code: "INTERNAL_ERROR"